+++ projects/ng2-smart-table/src/lib/lib/helpers.ts
@@ -1,14 +1,5 @@
 import { cloneDeep } from 'lodash';
 
-/**
- * Extending object that entered in first argument.
- *
- * Returns extended object or false if have no target object or incorrect type.
- *
- * If you wish to clone source object (without modify it), just use empty new
- * object as first argument, like this:
- *   deepExtend({}, yourObj_1, [yourObj_N]);
- */
 export const deepExtend = (...objects: Array<any>) => {
   if (!objects.length || typeof objects[0] !== 'object') {
     return false;
@@ -19,44 +10,33 @@ export const deepExtend = (...objects: Array<any>) => {
   }
 
   const target = objects[0];
-
-  // convert arguments to array and cut off target object
   const args = Array.prototype.slice.call(objects, 1);
 
-
   args.forEach((obj: any) => {
-    // skip argument if it is array or isn't object
     if (typeof obj !== 'object' || Array.isArray(obj)) {
       return;
     }
-
     Object.keys(obj).forEach((key) => {
-      const src = target[key]; // source value
-      const val = obj[key]; // new value
+      const src = target[key];
+      const val = obj[key];
 
-      // recursion prevention
       if (val === target) {
         return;
-
-      } else if (typeof val !== 'object' || val === null) {
+      }
+      if (typeof val !== 'object' || val === null) {
         target[key] = val;
         return;
-
-        // just clone arrays (and recursive clone objects inside)
-      } else if (Array.isArray(val)) {
+      }
+      if (Array.isArray(val)) {
         target[key] = cloneDeep(val);
         return;
-
-        // overwrite by new value if source isn't object or array
-      } else if (typeof src !== 'object' || src === null || Array.isArray(src)) {
+      }
+      if (typeof src !== 'object' || src === null || Array.isArray(src)) {
         target[key] = deepExtend({}, val);
         return;
-
-        // source value and new value is objects both, extending...
-      } else {
-        target[key] = deepExtend(src, val);
-        return;
       }
+      target[key] = deepExtend(src, val);
+      return;
     });
   });
 
@@ -78,14 +58,3 @@ export class Deferred {
   }
 }
 
-
-export function getDeepFromObject(object = {}, name: string, defaultValue?: any) {
-  const keys = name.split('.');
-  let level = deepExtend({}, object);
-  keys.forEach((k) => {
-    if (level && !!level[k]) {
-      level = level[k];
-    }
-  });
-  return !level ? defaultValue : level;
-}