+++ projects/ng2-smart-table/src/lib/ng2-smart-table.module.ts
@@ -9,6 +9,7 @@ import { TBodyModule } from './components/tbody/tbody.module';
 import { THeadModule } from './components/thead/thead.module';
 
 import { Ng2SmartTableComponent } from './ng2-smart-table.component';
+import { ColumnShowModule } from './components/columns-show/column-show.module';
 
 @NgModule({
   imports: [
@@ -20,6 +21,7 @@ import { Ng2SmartTableComponent } from './ng2-smart-table.component';
     PagerModule,
     TBodyModule,
     THeadModule,
+    ColumnShowModule
   ],
   declarations: [
     Ng2SmartTableComponent,